Separate GridMovement into 2 components
