+++ Assets/Scenes/secminhr_test_scene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 90b19908a2fa84b91a4f37b43f4c5f95
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
