+++ Assets/Scripts/Character/PointFollower.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bf848eb67cc74a5f905116e8aef9358
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
